Add paredit configuration and default forms
Introduce a new configuration module for paredit that includes type
definitions for pair forms and threading macros. This allows for
custom user configurations to extend the built-in defaults. The
default pair forms and threading macros are defined, enabling
flexible and customizable editing experiences.
